feat!: convert "manage create" args to flags (#25)
When running `penv manage create <foo>`, I've struggled to remember the
order for the CLI args to take. Converting them to flags allows them to
come in any order, as long as the arg of "ENVIRONMENT_ALIAS" is still
declared without a flag. Technically this is a breaking change to the
CLI, but I submit that's fine at this early stage, and provides a less
surprising interface.

Co-authored-by: Conor Schaefer <conor@penumbralabs.xyz>