fix(tests): update default penumbra version (#37)
* cleanup: docstring in cli integration tests

Submitting this as a PR to get the tests running and see what's broken.

* dev: use nextest in justfile

* ci: run integration tests on PRs

* fix(tests): update default penumbra version

The CLI changes for pclientd/pcli in v1.5.0 [0], so while penv does
support the new-style, it doesn't handle historical versions smartly.
Let's just bump the integration tests to use the latest, which is a
realistic scenario anyway.

[0] https://github.com/penumbra-zone/penumbra/releases/tag/v1.5.0